return by controller DTO/VoteTo

+++ src/main/java/com/github/SobolevRoman/restaurantvoting/service/VoteService.java
@@ -6,6 +6,7 @@ import com.github.SobolevRoman.restaurantvoting.model.User;
 import com.github.SobolevRoman.restaurantvoting.model.Vote;
 import com.github.SobolevRoman.restaurantvoting.repository.RestaurantRepository;
 import com.github.SobolevRoman.restaurantvoting.repository.VoteRepository;
+import com.github.SobolevRoman.restaurantvoting.to.VoteTo;
 import lombok.AllArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
@@ -14,6 +15,7 @@ import org.springframework.transaction.annotation.Transactional;
 import java.time.LocalDate;
 import java.util.List;
 import java.util.Optional;
+import java.util.stream.Collectors;
 
 import static com.github.SobolevRoman.restaurantvoting.util.validation.ValidationUtil.checkTime;
 
@@ -24,14 +26,18 @@ public class VoteService {
     private final VoteRepository voteRepository;
     private final RestaurantRepository restaurantRepository;
 
-    public List<Vote> getAll(int userId) {
+    public List<VoteTo> getAll(int userId) {
         log.info("get all vote");
-        return voteRepository.getAll(userId);
+        return voteRepository.getAll(userId).stream()
+                .map(vote -> new VoteTo(vote.id(), vote.getActualDate(), vote.getRestaurant().id()))
+                .collect(Collectors.toList());
     }
 
-    public Optional<Vote> get(int userId, int id) {
+    public Optional<VoteTo> get(int userId, int id) {
         log.info("get with id = {}", id);
-        return voteRepository.get(userId, id);
+        Vote v = voteRepository.get(userId, id).orElseThrow(
+                () -> new DataConflictException("Not found vote with id=" + id));
+        return Optional.of(new VoteTo(v.getId(), v.getActualDate(), v.getRestaurant().id()));
     }
 
     @Transactional
@@ -49,8 +55,10 @@ public class VoteService {
         checkTime();
         log.info("User {} update/change his mind for vote with id={}", user, id);
         Vote v = voteRepository.get(user.id(), id).orElseThrow(
-                () -> new DataConflictException("Not found vote today for User - " + user.getEmail())
-        );
+                () -> new DataConflictException("Not found vote with id=" + id));
+        if (v.getActualDate().compareTo(LocalDate.now()) != 0) {
+            throw new DataConflictException("Change mind for vote may be only for this date");
+        }
         voteRepository.delete(v);
     }
 }
